refactor pydantic models to dataclasses for internal stuff
