+++ audit-server/audit_server/log_storage/models.py
@@ -34,24 +34,8 @@ class UserLogEntry(pydantic.BaseModel):
 class UserLogForExport(pydantic.BaseModel):
     """User log data for export."""
 
-    user_log_entries: list[UserLogEntry]
+    userLogEntries: list[UserLogEntry]
     startedAt: datetime
     endedAt: Optional[datetime] = pydantic.Field(
         description="The time this period ended, or null if the period is still active."
     )
-
-
-class RobotLogPaths(pydantic.BaseModel):
-    """Robot Log Entry information to be exported."""
-
-    file_path: str
-    file_hash: str
-    file_sig: str
-    file_sig_version: str
-
-
-class LogPeriodEntries(pydantic.BaseModel):
-    """Entries associated with a particular log period."""
-
-    user_log: UserLogForExport
-    robot_log_entries: list[RobotLogPaths]
